ENH: Report along the way
