+++ docs/mkdocs.yml
@@ -59,8 +59,6 @@ nav:
             - BEM surface: settings/source/bem.md
             - Source space & forward solution: settings/source/forward.md
             - Inverse solution: settings/source/inverse.md
-        - Report:
-            - HTML Report: settings/report/report.md
     - Examples:
         - Examples Gallery: examples/examples.md
         - examples/ds003392.md
