--- docs/source/settings/report/report.md
@@ -1,3 +1,0 @@
-
-::: mne_bids_pipeline.config.report_evoked_n_time_points
-::: mne_bids_pipeline.config.report_stc_n_time_points
