+++ docs/source/settings/sensor/statistics.md
@@ -1,4 +1,5 @@
 ::: mne_bids_pipeline.config.contrasts
+::: mne_bids_pipeline.config.report_evoked_n_time_points
 ::: mne_bids_pipeline.config.decode
 ::: mne_bids_pipeline.config.decoding_metric
 ::: mne_bids_pipeline.config.decoding_n_splits
