+++ docs/source/settings/source/inverse.md
@@ -4,3 +4,4 @@
 ::: mne_bids_pipeline.config.noise_cov
 ::: mne_bids_pipeline.config.source_info_path_update
 ::: mne_bids_pipeline.config.inverse_targets
+::: mne_bids_pipeline.config.report_stc_n_time_points
