+++ mne_bids_pipeline/_config_utils.py
@@ -578,14 +578,12 @@ def _get_script_modules() -> Dict[str, Tuple[ModuleType]]:
     from .scripts import preprocessing
     from .scripts import sensor
     from .scripts import source
-    from .scripts import report
     from .scripts import freesurfer
 
     INIT_SCRIPTS = init.SCRIPTS
     PREPROCESSING_SCRIPTS = preprocessing.SCRIPTS
     SENSOR_SCRIPTS = sensor.SCRIPTS
     SOURCE_SCRIPTS = source.SCRIPTS
-    REPORT_SCRIPTS = report.SCRIPTS
     FREESURFER_SCRIPTS = freesurfer.SCRIPTS
 
     SCRIPT_MODULES = {
@@ -594,7 +592,6 @@ def _get_script_modules() -> Dict[str, Tuple[ModuleType]]:
         'preprocessing': PREPROCESSING_SCRIPTS,
         'sensor': SENSOR_SCRIPTS,
         'source': SOURCE_SCRIPTS,
-        'report': REPORT_SCRIPTS,
     }
 
     # Do not include the FreeSurfer scripts in "all" – we don't intend to run
@@ -603,8 +600,7 @@ def _get_script_modules() -> Dict[str, Tuple[ModuleType]]:
         SCRIPT_MODULES['init'] +
         SCRIPT_MODULES['preprocessing'] +
         SCRIPT_MODULES['sensor'] +
-        SCRIPT_MODULES['source'] +
-        SCRIPT_MODULES['report']
+        SCRIPT_MODULES['source']
     )
 
     return SCRIPT_MODULES
