--- mne_bids_pipeline/scripts/report/__init__.py
@@ -1,7 +1,0 @@
-"""Analysis report generation."""
-
-from . import _01_make_reports
-
-SCRIPTS = (
-    _01_make_reports,
-)
