+++ mne_bids_pipeline/scripts/source/_99_group_average.py
@@ -13,7 +13,8 @@ from mne_bids import BIDSPath
 
 from ..._config_utils import (
     get_fs_subjects_dir, get_subjects, sanitize_cond_name, get_fs_subject,
-    get_task, get_datatype, get_deriv_root, get_sessions, get_bids_root)
+    get_task, get_datatype, get_deriv_root, get_sessions, get_bids_root,
+    get_all_contrasts)
 from ..._logging import logger, gen_log_kwargs
 from ..._parallel import get_parallel_backend, parallel_func
 from ..._run import failsafe_run, save_logs
@@ -125,6 +126,8 @@ def get_config(
         exclude_subjects=config.exclude_subjects,
         sessions=get_sessions(config),
         use_template_mri=config.use_template_mri,
+        all_contrasts=get_all_contrasts(config),
+        report_stc_n_time_points=config.report_stc_n_time_points,
     )
     return cfg
