+++ mne_bids_pipeline/tests/run_tests.py
@@ -39,7 +39,7 @@ class TestOptionsT(TypedDict, total=False):
 # key: {
 #     'dataset': key.split('_')[0],
 #     'config': f'config_{key}.py',
-#     'steps': ('preprocessing', 'sensor', 'report'),
+#     'steps': ('preprocessing', 'sensor'),
 #     'env': {},
 #     'task': None,
 # }
@@ -57,13 +57,13 @@ TEST_SUITE: Dict[str, TestOptionsT] = {
     'ds000246': {
         'steps': ('preprocessing',
                   'preprocessing/make_epochs',  # Test the group/step syntax
-                  'sensor', 'report'),
+                  'sensor'),
     },
     'ds000247': {
         'task': 'rest',
     },
     'ds000248': {
-        'steps': ('preprocessing', 'sensor', 'source', 'report'),
+        'steps': ('preprocessing', 'sensor', 'source'),
     },
     'ds000248_ica': {},
     'ds000248_T1_BEM': {
@@ -76,13 +76,13 @@ TEST_SUITE: Dict[str, TestOptionsT] = {
         'steps': ('freesurfer/coreg_surfaces',),
     },
     'ds000248_no_mri': {
-        'steps': ('preprocessing', 'sensor', 'source', 'report'),
+        'steps': ('preprocessing', 'sensor', 'source'),
     },
     'ds001810': {
-        'steps': ('preprocessing', 'preprocessing', 'sensor', 'report'),
+        'steps': ('preprocessing', 'preprocessing', 'sensor'),
     },
     'ds003104': {
-        'steps': ('preprocessing', 'sensor',  'source', 'report'),
+        'steps': ('preprocessing', 'sensor',  'source'),
     },
     'ERP_CORE_N400': {
         'dataset': 'ERP_CORE',
@@ -181,7 +181,7 @@ def run_tests(test_suite, *, download, debug, cache):
 
         # Run the tests.
         steps = test_options.get(
-            'steps', ('preprocessing', 'sensor', 'report'))
+            'steps', ('preprocessing', 'sensor'))
         task = test_options.get('task', None)
         command = [
             'mne_bids_pipeline',
